UPSTREAM: sdm845: Implement bitbang UART for bootblock
This patch replaces the UART in the bootblock of SDM845 with a bitbang
implementation. Since SDM845 hardware UART needs a firmware blob loaded
into it before it becomes usable, it is not really suited for use in the
bootblock (since by the time we can read blobs from SPI, the bootblock
is essentially over anyway). This solution allows us to still have some
console output during early SoC initialization.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ia978154c052744fc4782aa6d3e2fb8dd740ed9ed
Signed-off-by: Martin Roth <martinroth@chromium.org>
Original-Commit-Id: 90f025b2762edbac0516cb42f4fc7c9c28f260fa
Original-Change-Id: I0c252ec83a7993edce5c4debc687f1fdd0d7b36d
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/25813
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1167897