UPSTREAM: mb/google/octopus: Enable pull on ESPI_IO1 line
This change configures a weak internal pull-up on ESPI_IO1 line for
octopus baseboard and variant bip. ESPI_IO1 is used as ALERT# line and
is expected to be open-drain. However, there is no external pull on
this line and so an internal pull-up is required to ensure proper eSPI
communication.

BUG=b:78497502
TEST=Verified that there is no eSPI communication failure between AP
and EC during boot-up and on suspend/resume.

Change-Id: I2346efdceee66147c0a9ef4c3297767d17aeeb05
Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Commit-Id: c0451791bf7ba3a378e05bb467cdb50939e0e8e0
Original-Change-Id: Ic494aa7397b94bfd233ce10da8287660997b3377
Original-Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-on: https://review.coreboot.org/25900
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Reviewed-by: Justin TerAvest <teravest@chromium.org>
Original-Reviewed-by: Hannah Williams <hannah.williams@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1034162
Commit-Ready: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>