UPSTREAM: chromeec platforms: Update ACPI thermal event handler call
Currently the thermal event handler method TEVT is defined as an extern,
then defined again in platforms with thermal event handling.  In newer
versions of IASL, this generates an error, as the method is defined in
two places.  Simply removing the extern causes the call to it to fail on
platforms where it isn't actually defined, so add a preprocessor define
where it's implemented, and only call the method on those platforms.

BUG=none
BRANCH=none
TEST=none

Change-Id: I02963a4f1647eeb95365befa13fa92c6b2f1345b
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 15f232df089a6620f7fe2541caa7713422885c72
Original-Change-Id: I64dcd2918d14f75ad3c356b321250bfa9d92c8a5
Original-Signed-off-by: Martin Roth <martinroth@google.com>
Original-Reviewed-on: https://review.coreboot.org/25916
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1039326
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>