UPSTREAM: software_i2c: Improve error behavior
This patch fixes a bug in the software_i2c (bitbang) framework where it
would previously not return an error when receiving a NACK on a write
transaction (deviating behavior from our hardware I2C drivers). It also
adds explicit error codes to be returned for the different kinds of
failure conditions so they are more useful for debugging when dumped.

BUG=none
BRANCH=none
TEST=none

Change-Id: I69c34502d124ceabe8b14dbae004f8517a4edb7f
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 6dc7b220a30aae322928cbb845988d9f5aa7ef78
Original-Change-Id: Ie63bf35123d89dcd99a1f9c079d4cae6a33b0b09
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/25963
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Reviewed-by: Jonathan Neuschfer <j.neuschaefer@gmx.net>
Reviewed-on: https://chromium-review.googlesource.com/1039338
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>