UPSTREAM: intel/broadwell: Add option to enable/disable the PCIe AER capability
The Advanced Error Reporting capability was hardcoded in the PCIe
extended capability list, but it might not always be possible.

The Librem 13v1 does not seem to have working AER and this option
was needed and tested on the Librem 13v1. Without it, the linux
console gets spammed with AER errrors.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ia8e0bbbb39b2eeae908fc6e322d92e764d61cdbf
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 71616784077cd5a712dcb8768cea3464c3de04ec
Original-Change-Id: If2e0ec42c93f1fee927eacdf0099004cf9302fbe
Original-Signed-off-by: Youness Alaoui <youness.alaoui@puri.sm>
Original-Reviewed-on: https://review.coreboot.org/25326
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1051592
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>