UPSTREAM: mb/lenovo/t520: Add Dual Graphics CMOS support
Adds the CMOS option for Dual Graphics, as is present in more Lenovo models
already. Enabling this option ensures that the NVIDIA GPU is powered.

More PCI devices can be observed when activating this setting. It was verified
on a W520, also by loading a VGA option ROM and achieving a working Dual
Graphics system.

The CMOS default has been kept to 'Integrated Only', as the usage of Dual
Graphics requires an option ROM and drains the battery more quickly.

BUG=none
BRANCH=none
TEST=none

Change-Id: I45ef4a853dbacf990f5cb77e9e36273757298de0
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 28cdd4c8d36c49cbe9b171a2af6077fa8a253498
Original-Change-Id: I41ccabd4554ca019684edd6f8b1c23679212c59f
Original-Signed-off-by: Nico Rikken <nico@nicorikken.eu>
Original-Reviewed-on: https://review.coreboot.org/26114
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Rudolph <siro@das-labor.org>
Reviewed-on: https://chromium-review.googlesource.com/1109792
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>