UPSTREAM: chromeec platforms: Update ACPI throttle handler call
Currently the throttle event handler method THRT is defined as an extern,
then defined again in the platform with thermal event handling.  In newer
versions of IASL, this generates an error, as the method is defined in
two places.  Simply removing the extern causes the call to it to fail on
platforms where it isn't actually defined, so add a preprocessor define
where it's implemented, and only call the method on those platforms.

BUG=none
BRANCH=none
TEST=none

Change-Id: Iac62c5c2a5d670edc0d3709a7d5ed10b0cc380b8
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 6b1ceacb9b30ec873b341811645413ae3e9e0331
Original-Change-Id: I6337c52edaf9350843848b31c5d87bbfca403930
Original-Signed-off-by: Martin Roth <martinroth@google.com>
Original-Reviewed-on: https://review.coreboot.org/26121
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1078430
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>