UPSTREAM: purism/librem_bdl: Convert to variant setup
Convert the purism/librem13v1 to a variant setup, in
preparation for adding the librem15v2 board as a new variant.
The Librem 13 v1 and Librem 15 v2 are nearly identical, so
this minimizes new code to add support for the latter.

Also update the URL in board_info to an archive.org link.

BUG=none
BRANCH=none
TEST=none

Change-Id: I1ae0a8c7a69723c8094b75b2f9ed577b660e02c1
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: b799e0df3d7d090884f5c1163804a17c93530599
Original-Change-Id: I00bb82b9e895e2464ddaa92915c01ce0e35933a2
Original-Signed-off-by: Youness Alaoui <youness.alaoui@puri.sm>
Original-Reviewed-on: https://review.coreboot.org/26183
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1055444
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>