UPSTREAM: mainboard/asus/am1i-a: turn on the tpm
Along with other patches submitted for review to get the chipset
parts working, this allows Linux or other OS to use a TPM module
plugged into the 20-pin LPC header on the board, by exposing its
presence through the ACPI and PNP tables.

This patch adds to the Kconfig and devicetree.cb files.

Tested with the TPM/FW 3.19 and the trousers tools.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ia34a3c32d15c2b340ae045bf3cb428e1f61e5d05
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: e36a00af719347f1f26d37d4b13fe2d416cdfef1
Original-Change-Id: I8c1aea245f81fa44a6bdd5301bbee958cbcdfaaa
Original-Signed-off-by: Kevin Cody-Little <kcodyjr@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/26193
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://chromium-review.googlesource.com/1084716
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>