UPSTREAM: cbfs_locate: Optionally return file type
In some cases callers want to know if a file
exists and, if so, what its type is.

Modify cbfs_locate so that if the pointer is non-NULL,
but has the value 0, the type of the file that
matches the name will be returned.

BUG=none
BRANCH=none
TEST=none

Change-Id: I3acc7a8dc6f08b4cc4416fecdb586e52a6be44c4
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 60fd684698e7da894d539936307aea94d0c83bd1
Original-Change-Id: Ic1349d358c3054207ccbccf3825db56784327ad0
Original-Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/26279
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1059475
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>