UPSTREAM: lib/gnat: Drop Restriction `No_Exception_Propagation`
It turns out that even with the `-gnatp` switch to suppress runtime
checks, the compiler is still allowed to generate them (it only doesn't
have to). If we can't control generation of checks, we also can't
make assumptions about propagation of their exceptions.

The compiler warning that led to this change seems spurious, though
(the check might be generated, but is dropped later). So we might
revert this decision if the compiler can be fixed.

BUG=none
BRANCH=none
TEST=none

Change-Id: Id5b43ab522e3a01999709b30c804b253a2ae762a
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 4a027e6e95665f2c7cb9297c04d60e1b58b815b2
Original-Change-Id: I7470d74b1f96f90d0d15b24dfd636d5f1c778d46
Original-Signed-off-by: Nico Huber <nico.huber@secunet.com>
Original-Reviewed-on: https://review.coreboot.org/26350
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1068720
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>