UPSTREAM: nb/intel/nehalem: Fix smashed stack in romstage
Stack smashing was detected during raminit when not loading from MRC.
Adding CAR_GLOBAL to a struct inside raminit was suggested in
https://mail.coreboot.org/pipermail/coreboot/2018-May/086677.html in
order to fix the problem.
Adding CAR_GLOBAL to the ram timings variable solves the issue (adding
it to the ram_training or raminfo struct had no effect).
This is just a workaround and might need a proper fix in the future.

Tested on Lenovo X201i with 2+2 and 4+4 GB RAM.

BUG=none
BRANCH=none
TEST=none

Change-Id: Id7d67d24263e32d957943cf67943758a7b1aec51
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: dfa51259ad1434aa75bef6d071ba5dfacbf13cba
Original-Change-Id: I21b380db61be2aedc045201821d83e18e7d07ad1
Original-Signed-off-by: Matthias Gazzari <mail@qtux.eu>
Original-Reviewed-on: https://review.coreboot.org/26388
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Kysti Mlkki <kyosti.malkki@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/1068813
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>